chore: Rename Cimple.Analysis to Linter.
The "Cimple" used to be useful because we used to have "C" as well.
"Analysis" will be used for analyses that are then used by linters.
Linters themselves do analyse things, but are more specific. Analysers
will be modules that do generic things like liveness analysis.
