+++ src/Tokstyle/Linter/ForLoops.hs
@@ -1,9 +1,11 @@
 {-# LANGUAGE OverloadedStrings #-}
-module Tokstyle.Cimple.Analysis.FuncPrototypes (analyse) where
+module Tokstyle.Linter.ForLoops (analyse) where
 
 import qualified Control.Monad.State.Lazy    as State
 import           Data.Text                   (Text)
-import           Language.Cimple             (AstActions, Lexeme, Node (..),
+import qualified Data.Text                   as Text
+import           Language.Cimple             (AssignOp (..), AstActions,
+                                              Lexeme (..), Node (..),
                                               defaultActions, doNode,
                                               traverseAst)
 import qualified Language.Cimple.Diagnostics as Diagnostics
@@ -13,12 +15,15 @@ linter :: AstActions [Text]
 linter = defaultActions
     { doNode = \file node act ->
         case node of
-            FunctionPrototype _ name [] -> do
-                Diagnostics.warn file name "empty parameter list must be written as (void)"
+            ForStmt (VarDecl _ty (Declarator (DeclSpecVar _i) (Just _v))) _ _ _ -> act
+            ForStmt (AssignExpr (VarExpr _i) AopEq _v) _ _ _ -> act
+            ForStmt i _ _ _ -> do
+                warn file node . Text.pack . show $ i
                 act
 
             _ -> act
     }
+  where warn file node = Diagnostics.warn file (Diagnostics.at node)
 
 analyse :: (FilePath, [Node () (Lexeme Text)]) -> [Text]
 analyse = reverse . flip State.execState [] . traverseAst linter
