+++ tokstyle.cabal
@@ -18,18 +18,18 @@ source-repository head
 library
   default-language:    Haskell2010
   exposed-modules:
-      Tokstyle.Cimple.Analysis
+      Tokstyle.Linter
   other-modules:
-      Tokstyle.Cimple.Analysis.DeclaredOnce
-    , Tokstyle.Cimple.Analysis.DeclsHaveDefns
-    , Tokstyle.Cimple.Analysis.DocComments
-    , Tokstyle.Cimple.Analysis.ForLoops
-    , Tokstyle.Cimple.Analysis.FuncPrototypes
-    , Tokstyle.Cimple.Analysis.FuncScopes
-    , Tokstyle.Cimple.Analysis.GlobalFuncs
-    , Tokstyle.Cimple.Analysis.LoggerCalls
-    , Tokstyle.Cimple.Analysis.LoggerNoEscapes
-    , Tokstyle.Cimple.Analysis.VarUnusedInScope
+      Tokstyle.Linter.DeclaredOnce
+    , Tokstyle.Linter.DeclsHaveDefns
+    , Tokstyle.Linter.DocComments
+    , Tokstyle.Linter.ForLoops
+    , Tokstyle.Linter.FuncPrototypes
+    , Tokstyle.Linter.FuncScopes
+    , Tokstyle.Linter.GlobalFuncs
+    , Tokstyle.Linter.LoggerCalls
+    , Tokstyle.Linter.LoggerNoEscapes
+    , Tokstyle.Linter.VarUnusedInScope
     , Tokstyle.Result
   ghc-options:
       -Wall
@@ -86,7 +86,7 @@ test-suite testsuite
   hs-source-dirs: test
   main-is: testsuite.hs
   other-modules:
-      Tokstyle.Cimple.AnalysisSpec
+      Tokstyle.LinterSpec
   ghc-options:
       -Wall
       -fno-warn-unused-imports
